Refactor: add useCoffeeSelection hook for use favorite menu info data in registration process and refactor coffeeSelection componenet
